+++ SmallBeachFarm/docs/release-notes.md
@@ -2,6 +2,8 @@
 
 # Release notes
 ## Upcoming option
+* Added 'spawn monsters by default' config option (default false).  
+  _This sets the default value for the 'spawn monsters at night' option when creating a new save. This has no effect on existing saves (see [how to toggle it for an existing save](https://stardewvalleywiki.com/Monsters#Monsters_on_The_Farm))._
 * Added sections in config UI.
 
 ## 2.2.0